chore: rename the nav menu components to be apt and legible
